server_support [nfc]: Move some server-support logic out to this new file
